PG-033 ensure concurrency-safe state transition

+++ src/main/java/com/example/rankinggame/repositories/AnswerRepository.java
@@ -10,7 +10,7 @@ public interface AnswerRepository {
 
     boolean existsByRoundIdAndPlayerId(UUID roundId, UUID playerId);
 
-    long countByRoundId(UUID roundId);
+    int countByRoundId(UUID roundId);
 
     List<AnswerEntity> findByRoundIdOrderBySubmittedAtAsc(UUID roundId);
 }
